pyproject.toml
[tool.poetry]
name = "showbay-backend-assessment"
version = "0.1.0"
description = "Python Backend Engineer Assessment - FastAPI with PostgreSQL"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlmodel = "^0.0.16"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
sqlalchemy = "^2.0.23"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.12.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["showbay/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"